+++ lib/src/generator/representation.dart
@@ -11,56 +11,72 @@ String representation({
   var indent = initialIndent;
   final buffer = stringBuffer ??= StringBuffer();
   final nullableSymbol = nullable ? '?' : '';
+
   void innerRepresent(Object? object) {
-    if (object is Iterable<Object?>) {
-      buffer
-        ..write('<Object')
-        ..write(nullableSymbol)
-        ..writeln('>[');
-      indent++;
-      for (final entry in object) {
-        buffer.write(' ' * indent * 2);
-        innerRepresent(entry);
-        buffer.writeln(',');
-      }
-      indent--;
-      buffer
-        ..write(' ' * indent * 2)
-        ..write(']');
-    } else if (object is Map<String, Object?>) {
-      buffer
-        ..write('<String, Object')
-        ..write(nullableSymbol)
-        ..writeln('>{');
-      indent++;
-      for (final entry in object.entries) {
+    switch (object) {
+      case Iterable<Object?> iterable:
         buffer
-          ..write(' ' * indent * 2)
-          ..write('\'')
-          ..write(entry.key)
-          ..write('\'')
-          ..write(': ');
-        innerRepresent(entry.value);
-        buffer.writeln(',');
-      }
-      indent--;
-      buffer
-        ..write(' ' * indent * 2)
-        ..write('}');
-    } else if (object is String) {
-      buffer
-        ..write('r\'')
-        ..write(object.replaceAll('\r', r'\r').replaceAll('\n', r'\n'))
-        ..write('\'');
-    } else if (object is num || object is bool) {
-      buffer.write(object.toString());
-    } else if (object == null) {
-      buffer.write('null');
-    } else {
-      innerRepresent(object.toString());
+          ..write('<Object')
+          ..write(nullableSymbol)
+          ..writeln('>[');
+        indent++;
+        for (final entry in iterable) {
+          buffer.write('  ' * indent);
+          innerRepresent(entry);
+          buffer.writeln(',');
+        }
+        indent--;
+        buffer
+          ..write('  ' * indent)
+          ..write(']');
+
+      case Map<String, Object?> map:
+        buffer
+          ..write('<String, Object')
+          ..write(nullableSymbol)
+          ..writeln('>{');
+        indent++;
+        for (final entry in map.entries) {
+          buffer
+            ..write('  ' * indent)
+            ..write('\'')
+            ..write(_escapeString(entry.key))
+            ..write('\': ');
+          innerRepresent(entry.value);
+          buffer.writeln(',');
+        }
+        indent--;
+        buffer
+          ..write('  ' * indent)
+          ..write('}');
+
+      case String str:
+        buffer
+          ..write('r\'')
+          ..write(_escapeString(str))
+          ..write('\'');
+
+      case num number:
+        buffer.write(number.toString());
+
+      case bool boolean:
+        buffer.write(boolean.toString());
+
+      case null:
+        buffer.write('null');
+
+      default:
+        innerRepresent(object.toString());
     }
   }
 
   innerRepresent(source);
   return buffer.toString();
 }
+
+String _escapeString(String input) => input
+    .replaceAll(r'\', r'\\')
+    .replaceAll('\'', '\\\'')
+    .replaceAll('\r', r'\r')
+    .replaceAll('\n', r'\n')
+    .replaceAll('\t', r'\t');